Add Player and Camera classes, update shaders
In this commit, we've added a Player class that uses keyboard and mouse controls, and a Camera class for viewing calculations. We've also updated the shader_program.py to include these changes. An additional update has been made to the quad.vert shader file to use the provided matrices for projecting, viewing and modeling.